Adds second method of analysis
At this point I have a pretty good idea what an 'analyzer' needs to be.
I'm not sure if I need an abstraction over the alerts it provides or if
I should just go with a single SpendingAlert class.

The main thing I don't like is I know I will have this magic string in
both that really identifies what the alert is for(clothing/category,
monthly income) and then the analyzer will have to know too much about
the formatting of an alert message. Another option would be some sort of
enum but that still seems like it will be too coupled.

Going to defer that decisions for the time being and just leave it as-is
while working on extracting analyzer behavior