Adding some structure
Now that I have a better idea of what the pieces are going to be I can
apply some structure to it.

One thing that really jumped out was the inconsistency with the term
'Spending'
Sometimes it was referring to the act of spending money and other times
it applied to your budget.  I wanted to clear that up so I renamed all
the stuff that relates to you spending money is under Activity.
Right now I just have a report with some financial transactions (because
transaction by itself is such a loaded word with computers)

I want to clean up my BudgetTool api a little bit and then add support
for loading from files so i can get my integration test passing again.

Then I need to decide what I want to do with alerts